feat(Option): 옵션을 product에서 분리함
